omninwm/models/OmniNWM-VLA/requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
archspec==0.2.5
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
autoroot==1.0.1
av==14.2.0
beautifulsoup4==4.13.5
boltons==25.0.0
brotli==1.1.0
causal-conv1d==1.5.0.post8
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datasets==3.4.1
decorator==5.2.1
decord==0.6.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
dropout-layer-norm==0.1
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
expecttest==0.2.1
filelock==3.13.1
flash-attn==2.7.4.post1
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
fused-dense-lib==0.0.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.29.1
hyperframe==6.1.0
hypothesis==6.115.5
idna==3.10
importlib-resources==6.5.2
ipython==9.5.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
libarchive-c==5.3
lief==0.17.0
lintrunner==0.12.5
mamba-ssm==2.2.4
markupsafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.1
numerize==0.12
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
optree==0.13.0
packaging==25.0
pandas==2.2.3
parso==0.8.5
peft==0.14.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pillow-heif==0.21.0
pkginfo==1.12.1.2
pkgutil-resolve-name==1.3.10
platformdirs==4.4.0
pluggy==1.6.0
prompt-toolkit==3.0.52
propcache==0.3.0
protobuf==5.29.3
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pycosat==0.6.6
pycparser==2.23
pygments==2.19.2
pyparsing==3.2.1
pyrootutils==1.0.4
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-etcd==0.4.5
pytz==2025.2
pyyaml==6.0.3
qwen-vl-utils==0.0.10
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
rotary-emb==0.1
rpds-py==0.27.1
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.14
safetensors==0.5.3
sentencepiece==0.2.0
shutup==0.2.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.13.1
timm==1.0.15
tokenizers==0.21.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchelastic==0.2.2
torchvision==0.20.1+cu124
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.1.0
truststore==0.10.4
types-dataclasses==0.6.6
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.5.0
wcwidth==0.2.14
xentropy-cuda-lib==0.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.23.0
zstandard==0.23.0